requirements2.txt
flask
flask-cors
mysql-connector-python
bcrypt
pyjwt
# scikit-learn==1.4.2
pandas
numpy
openai
python-dotenv
pytest

reportlab==4.0.4
xhtml2pdf==0.2.11
Jinja2==3.1.2

# pip install scikit-learn==1.5.2
# pip install numpy==2.1.3 pandas==2.2.3 scipy==1.13.1
# python -m pip install --upgrade pip setuptools wheel
# pip install PyJWT 
# pip install reportlab xhtml2pdf